Revert "made the application work and removed some of the erros in the code"
