+++ .env
@@ -1,0 +1,3 @@
+VITE_SUPABASE_PROJECT_ID="gkegdnhblpaccdnwwlba"
+VITE_SUPABASE_PUBLISHABLE_KEY="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImdrZWdkbmhibHBhY2Nkbnd3bGJhIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NTY2NTkzMDgsImV4cCI6MjA3MjIzNTMwOH0.R4Anb2bBZFDEA5cfSM0PkYdAHrbRhSmwCcFac-oX-6Y"
+VITE_SUPABASE_URL="https://gkegdnhblpaccdnwwlba.supabase.co"
