+++ .gitignore
@@ -12,8 +12,6 @@ dist
 dist-ssr
 *.local
 
-.env
-
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
