+++ src/components/FarmerDashboard.tsx
@@ -9,7 +9,6 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@
 import { useToast } from '@/hooks/use-toast';
 import { Plus, Package, TrendingUp, Users } from 'lucide-react';
 import { Badge } from '@/components/ui/badge';
-import {Header} from "@/components/Header";
 
 interface Profile {
   id: string;
@@ -37,11 +36,8 @@ interface SurplusListing {
   created_at: string;
 }
 
-interface FarmerDashboardProps {
-  profile: Profile;
-}
-
-export const FarmerDashboard: React.FC<FarmerDashboardProps> = ({profile}) => {
+export const FarmerDashboard: React.FC = () => {
+  const [profile, setProfile] = useState<Profile | null>(null);
   const [listings, setListings] = useState<SurplusListing[]>([]);
   const [showAddForm, setShowAddForm] = useState(false);
   const [loading, setLoading] = useState(false);
@@ -76,6 +72,8 @@ export const FarmerDashboard: React.FC<FarmerDashboardProps> = ({profile}) => {
         });
         return;
       }
+
+      setProfile(profilesData);
 
       // Fetch listings
       await fetchListings(session.user.id);
@@ -149,14 +147,10 @@ export const FarmerDashboard: React.FC<FarmerDashboardProps> = ({profile}) => {
 
       // Trigger AI matching
       await triggerAIMatching(profile.user_id);
-    } catch (error: unknown) {
-      let errorMessage = "An unexpected error occurred";
-      if (error instanceof Error) {
-        errorMessage = error.message;
-      }
+    } catch (error: any) {
       toast({
         title: "Error",
-        description: errorMessage,
+        description: error.message,
         variant: "destructive"
       });
     } finally {
@@ -194,9 +188,6 @@ export const FarmerDashboard: React.FC<FarmerDashboardProps> = ({profile}) => {
 
   return (
     <div className="space-y-6">
-      <Header />
-      <h1 className="text-3xl font-bold">Welcome, {profile.full_name}</h1>
-      <p className="text-muted-foreground">Manage your surplus listings and connect with community kitchens</p>
       {/* Stats Cards */}
       <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
         <Card>
