+++ src/pages/Index.tsx
@@ -1,5 +1,5 @@
 import { Header } from "@/components/Header";
-import { Home } from "@/components/Home";
+import { Hero } from "@/components/Hero";
 import { Features } from "@/components/Features";
 import { HowItWorks } from "@/components/HowItWorks";
 import { CTA } from "@/components/CTA";
@@ -8,7 +8,7 @@ const Index = () => {
   return (
     <div className="min-h-screen">
       <Header />
-      <Home />
+      <Hero />
       <Features />
       <HowItWorks />
       <CTA />
